requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
cupy-cuda12x==13.4.1
cython==3.1.2
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fastrlock==0.8.3
ipykernel==6.29.5
ipython==9.3.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jupyter-client==8.6.3
jupyter-core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
parso==0.8.4
pip==25.1.1
platformdirs==4.3.8
prompt-toolkit==3.0.51
psutil==7.0.0
pure-eval==0.2.3
pygments==2.19.1
python-dateutil==2.9.0.post0
pywin32==310
pyzmq==26.4.0
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
wcwidth==0.2.13
